Change cmpBuffs' arguments type to uin8_t*

+++ modules/utils.c
@@ -72,7 +72,7 @@ uint8_t calculate_chksum ( uint8_t * buffer, uint8_t range )
 
 /* Compare two buffers' size and data
  * Returns 0 if equal, 0xFF if different */
-uint8_t cmpBuffs( uint32_t *bufa, uint32_t len_a, uint32_t *bufb, uint32_t len_b )
+uint8_t cmpBuffs( uint8_t *bufa, size_t len_a, uint8_t *bufb, size_t len_b )
 {
     uint16_t i;
     if (len_a != len_b) {
